Node Prep
Setting up INode class to allow for nodes to be added.

Reworked getCenter() to allow for normal text and also special text such
as text with icons included in them.

Fixed bug with drawRectWithUV() not rendering the correct texture when
using UtilFatal texture

Removed jsonToClass as of right now to think of a much more cleaner and
beneficial way to read and write json files.